+++ packages/expo-video/android/src/main/java/expo/modules/video/records/VideoMetadata.kt
@@ -1,0 +1,10 @@
+package expo.modules.video.records
+
+import expo.modules.kotlin.records.Field
+import expo.modules.kotlin.records.Record
+import java.io.Serializable
+
+class VideoMetadata(
+  @Field var title: String? = null,
+  @Field var artist: String? = null
+) : Record, Serializable